Avtfm 97

+++ api/activities.js
@@ -5,7 +5,6 @@
  *      date, // Date of the activity
  *      clientId,
  *      createdByUserId, // id of the user which created the activity. Only this user can change the activity
- *      participantUserIds, // List of user ids of participants of this activity. Those users can see the activity
  *      name, // Short name for the activity list
  *      task, // Detailed description of the tasks to be done
  *      isDone, // TRUE / FALSE whether the tasks were completed
@@ -23,8 +22,7 @@ var co = require('../utils/constants');
 var rh = require('../utils/relationsHelper');
 var dah = require('../utils/dynamicAttributesHelper');
 
-// TODO: Mechanismus überlegen, der die createdByUserId und participantUserIdsin richtige ObjectIDs umwandelt
-// Get all activities of the current client, which were created by the current user or where the current user is a participant of
+// Get all activities of the current client, which were created by the current user or which are public
 router.get('/', auth(co.permissions.OFFICE_ACTIVITY, 'r', co.modules.activities), (req, res) => {
     var userId = req.user._id;
     var clientId = req.user.clientId; // clientId === null means that the user is a portal user
@@ -32,7 +30,7 @@ router.get('/', auth(co.permissions.OFFICE_ACTIVITY, 'r', co.modules.activities)
         clientId: clientId,
         $or: [
             { createdByUserId: userId },
-            { participantUserIds: userId } // Array query, see https://docs.mongodb.com/manual/tutorial/query-arrays/#query-an-array-for-an-element
+            { isForAllUsers: true } // Array query, see https://docs.mongodb.com/manual/tutorial/query-arrays/#query-an-array-for-an-element
         ]
     };
     req.db.get(co.collections.activities.name).find(query, req.query.fields).then((activities) => {
@@ -63,9 +61,9 @@ router.get('/forIds', auth(false, false, co.modules.activities), (req, res) => {
         req.db.get(co.collections.activities.name).find({
             _id: { $in: ids },
             clientId: clientId,
-            $or: [ // Nur die Termine, die der Benutzer auch angelegt hat
+            $or: [ // Nur die Termine, die der Benutzer auch angelegt hat oder die öffentlich sind
                 { createdByUserId: userId },
-                { participantUserIds: userId } 
+                { isForAllUsers: true } 
             ]
         }).then((activities) => {
             res.send(activities);
@@ -76,9 +74,16 @@ router.get('/forIds', auth(false, false, co.modules.activities), (req, res) => {
 // Get a specific activity
 router.get('/:id', auth(co.permissions.OFFICE_ACTIVITY, 'r', co.modules.activities), validateId, validateSameClientId(co.collections.activities.name), (req, res) => {
     req.db.get(co.collections.activities.name).findOne(req.params.id, req.query.fields).then((activity) => {
+        if (!activity.isForAllUsers && !req.user._id.equals(activity.createdByUserId)) {
+            res.sendStatus(403); // Zugriff nur auf eigene oder auf öffentliche Termine möglich
+            return;
+        }
         // Database element is available here in every case, because validateSameClientId already checked for existence
-        activity.fullyEditable = activity.createdByUserId.equals(req.user._id); // Flag to show whether the client can edit all properties or only isDone and comment
-        res.send(activity);
+        activity.currentUserCanWrite = activity.createdByUserId.equals(req.user._id); // Flag to show whether the client can edit all properties or only isDone and comment
+        req.db.get(co.collections.users.name).findOne(activity.createdByUserId).then((creator) => {
+            activity.creator = creator.name;
+            res.send(activity);
+        });
     });
 });
 
@@ -91,12 +96,6 @@ router.post('/', auth(co.permissions.OFFICE_ACTIVITY, 'w', co.modules.activities
     delete activity._id; // Ids are generated automatically
     activity.clientId = req.user.clientId;
     activity.createdByUserId = req.user._id; // Define the current user as the creator of the activity
-    // Teilnehmer-IDs umwandeln, falls vorhanden, die Existenz wird dabei nicht geprüft
-    if (activity.participantUserIds) {
-        activity.participantUserIds = activity.participantUserIds.filter(validateId.validateId).map(function(id) { return monk.id(id); });
-    } else {
-        activity.participantUserIds = [];
-    }
     req.db.insert(co.collections.activities.name, activity).then((insertedActivity) => {
         res.send(insertedActivity);
     });
@@ -115,17 +114,8 @@ router.put('/:id', auth(co.permissions.OFFICE_ACTIVITY, 'w', co.modules.activiti
     req.db.get(co.collections.activities.name).findOne(req.params.id).then((existingActivity) => {
         // Database element is available here in every case, because validateSameClientId already checked for existence
         // Forbid the change of specific activity data when the current user is not the creator
-        if (!existingActivity.createdByUserId.equals(req.user._id) && (
-            activity.date ||
-            activity.participantUserIds ||
-            activity.task ||
-            activity.type
-        )) {
+        if (!existingActivity.createdByUserId.equals(req.user._id)) {
             return res.sendStatus(403); // Forbidden
-        }
-        // Teilnehmer-IDs umwandeln, falls vorhanden
-        if (activity.participantUserIds) {
-            activity.participantUserIds = activity.participantUserIds.filter(validateId.validateId).map(function(id) { return monk.id(id); });
         }
         req.db.update(co.collections.activities.name, req.params.id, { $set: activity }).then((updatedActivity) => { // https://docs.mongodb.com/manual/reference/operator/update/set/
             res.send(updatedActivity);
@@ -138,12 +128,17 @@ router.put('/:id', auth(co.permissions.OFFICE_ACTIVITY, 'w', co.modules.activiti
  */
 router.delete('/:id', auth(co.permissions.OFFICE_ACTIVITY, 'w', co.modules.activities), validateId, validateSameClientId(co.collections.activities.name), function(req, res) {
     var activityId = monk.id(req.params.id);
-    req.db.remove(co.collections.activities.name, activityId).then((result) => {
-        return rh.deleteAllRelationsForEntity(co.collections.activities.name, activityId);
-    }).then(() => {
-        return dah.deleteAllDynamicAttributeValuesForEntity(activityId);
-    }).then(() => {
-        res.sendStatus(204);
+    req.db.get(co.collections.activities.name).findOne(activityId).then((existingActivity) => {
+        if (!existingActivity.createdByUserId.equals(req.user._id)) { // Es können nur eigene Termine gelöscht werden
+            return res.sendStatus(403); // Forbidden
+        }
+        req.db.remove(co.collections.activities.name, activityId).then((result) => {
+            return rh.deleteAllRelationsForEntity(co.collections.activities.name, activityId);
+        }).then(() => {
+            return dah.deleteAllDynamicAttributeValuesForEntity(activityId);
+        }).then(() => {
+            res.sendStatus(204);
+        });
     });
 });
